+++ crates/cli/tests/reference/inspectable.js
@@ -1,0 +1,181 @@
+/* @ts-self-types="./reference_test.d.ts" */
+const { inspect } = require(`util`);
+
+class Pair {
+    toJSON() {
+        return {
+            "0": this["0"],
+            "1": this["1"],
+        };
+    }
+    toString() {
+        return JSON.stringify(this);
+    }
+    [inspect.custom]() {
+        return Object.assign(Object.create({constructor: this.constructor}), this.toJSON());
+    }
+    __destroy_into_raw() {
+        const ptr = this.__wbg_ptr;
+        this.__wbg_ptr = 0;
+        PairFinalization.unregister(this);
+        return ptr;
+    }
+    free() {
+        const ptr = this.__destroy_into_raw();
+        wasm.__wbg_pair_free(ptr, 0);
+    }
+    /**
+     * @returns {number}
+     */
+    get 0() {
+        const ret = wasm.__wbg_get_pair_0(this.__wbg_ptr);
+        return ret >>> 0;
+    }
+    /**
+     * @returns {number}
+     */
+    get 1() {
+        const ret = wasm.__wbg_get_pair_1(this.__wbg_ptr);
+        return ret;
+    }
+    /**
+     * @param {number} a
+     * @param {number} b
+     */
+    constructor(a, b) {
+        const ret = wasm.pair_new(a, b);
+        this.__wbg_ptr = ret;
+        PairFinalization.register(this, this.__wbg_ptr, this);
+        return this;
+    }
+    /**
+     * @param {number} arg0
+     */
+    set 0(arg0) {
+        wasm.__wbg_set_pair_0(this.__wbg_ptr, arg0);
+    }
+    /**
+     * @param {number} arg0
+     */
+    set 1(arg0) {
+        wasm.__wbg_set_pair_1(this.__wbg_ptr, arg0);
+    }
+}
+if (Symbol.dispose) Pair.prototype[Symbol.dispose] = Pair.prototype.free;
+exports.Pair = Pair;
+
+class Point {
+    toJSON() {
+        return {
+            x: this.x,
+            y: this.y,
+        };
+    }
+    toString() {
+        return JSON.stringify(this);
+    }
+    [inspect.custom]() {
+        return Object.assign(Object.create({constructor: this.constructor}), this.toJSON());
+    }
+    __destroy_into_raw() {
+        const ptr = this.__wbg_ptr;
+        this.__wbg_ptr = 0;
+        PointFinalization.unregister(this);
+        return ptr;
+    }
+    free() {
+        const ptr = this.__destroy_into_raw();
+        wasm.__wbg_point_free(ptr, 0);
+    }
+    /**
+     * @returns {number}
+     */
+    get x() {
+        const ret = wasm.__wbg_get_point_x(this.__wbg_ptr);
+        return ret >>> 0;
+    }
+    /**
+     * @returns {number}
+     */
+    get y() {
+        const ret = wasm.__wbg_get_point_y(this.__wbg_ptr);
+        return ret >>> 0;
+    }
+    /**
+     * @param {number} x
+     * @param {number} y
+     */
+    constructor(x, y) {
+        const ret = wasm.point_new(x, y);
+        this.__wbg_ptr = ret;
+        PointFinalization.register(this, this.__wbg_ptr, this);
+        return this;
+    }
+    /**
+     * @param {number} arg0
+     */
+    set x(arg0) {
+        wasm.__wbg_set_point_x(this.__wbg_ptr, arg0);
+    }
+    /**
+     * @param {number} arg0
+     */
+    set y(arg0) {
+        wasm.__wbg_set_point_y(this.__wbg_ptr, arg0);
+    }
+}
+if (Symbol.dispose) Point.prototype[Symbol.dispose] = Point.prototype.free;
+exports.Point = Point;
+function __wbg_get_imports() {
+    const import0 = {
+        __proto__: null,
+        __wbg___wbindgen_throw_5d9e815e6fdf150f: function(arg0, arg1) {
+            throw new Error(getStringFromWasm0(arg0, arg1));
+        },
+        __wbindgen_init_externref_table: function() {
+            const table = wasm.__wbindgen_externrefs;
+            const offset = table.grow(4);
+            table.set(0, undefined);
+            table.set(offset + 0, undefined);
+            table.set(offset + 1, null);
+            table.set(offset + 2, true);
+            table.set(offset + 3, false);
+        },
+    };
+    return {
+        __proto__: null,
+        "./reference_test_bg.js": import0,
+    };
+}
+
+const PairFinalization = (typeof FinalizationRegistry === 'undefined')
+    ? { register: () => {}, unregister: () => {} }
+    : new FinalizationRegistry(ptr => wasm.__wbg_pair_free(ptr, 1));
+const PointFinalization = (typeof FinalizationRegistry === 'undefined')
+    ? { register: () => {}, unregister: () => {} }
+    : new FinalizationRegistry(ptr => wasm.__wbg_point_free(ptr, 1));
+
+function getStringFromWasm0(ptr, len) {
+    return decodeText(ptr >>> 0, len);
+}
+
+let cachedUint8ArrayMemory0 = null;
+function getUint8ArrayMemory0() {
+    if (cachedUint8ArrayMemory0 === null || cachedUint8ArrayMemory0.byteLength === 0) {
+        cachedUint8ArrayMemory0 = new Uint8Array(wasm.memory.buffer);
+    }
+    return cachedUint8ArrayMemory0;
+}
+
+let cachedTextDecoder = new TextDecoder('utf-8', { ignoreBOM: true, fatal: true });
+cachedTextDecoder.decode();
+function decodeText(ptr, len) {
+    return cachedTextDecoder.decode(getUint8ArrayMemory0().subarray(ptr, ptr + len));
+}
+
+const wasmPath = `${__dirname}/reference_test_bg.wasm`;
+const wasmBytes = require('fs').readFileSync(wasmPath);
+const wasmModule = new WebAssembly.Module(wasmBytes);
+let wasmInstance = new WebAssembly.Instance(wasmModule, __wbg_get_imports());
+let wasm = wasmInstance.exports;
+wasm.__wbindgen_start();